fix: generating id while putting entry in db

+++ EmployeeManagementSystem/src/main/java/com/example/EmployeeManagementSystem/dto/DtoEmployee.java
@@ -1,7 +1,7 @@
 package com.example.EmployeeManagementSystem.dto;
 
 public class DtoEmployee {
-    private int id;
+    private Long id;
     private String firstname;
     private String lastname;
     private String email;
@@ -9,18 +9,18 @@ public class DtoEmployee {
     public DtoEmployee() {
     }
 
-    public DtoEmployee(int id, String firstname, String lastname, String email) {
+    public DtoEmployee(Long id, String firstname, String lastname, String email) {
         this.id = id;
         this.firstname = firstname;
         this.lastname = lastname;
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
